marketdata_token_service/requirements.txt
requests
PyJWT
backoff
pytest
pytest-mock